feat: sistema de impuestos por país (IVA/VAT) + reglas de margen de ganancia
Backend - Country Taxes:
- Liquibase migration db.changelog-3.4.sql: tabla country_taxes con 41 registros (33 estándar + 8 reducidos)
- CountryTax domain model, entity, repository, JPA, infra mapper
- CountryTaxUseCase + impl: CRUD completo
- TaxCalculationService: cálculo por país/región con resolución jerárquica
- CountryTaxController: REST /api/v1/taxes con endpoint /calculate
- DTOs: CountryTaxDtoIn, CountryTaxDtoOut, TaxCalculationDtoOut
- CountryTaxApiMapper: MapStruct con conversión appliesTo<->appliesToCategories

Backend - Price Rules:
- Liquibase migrations db.changelog-3.2.sql + 3.3.sql: tabla price_rules con rangos
- PriceRule domain model, entity, repository, JPA, infra mapper
- PriceRuleUseCase + impl: CRUD con invalidación de cache
- PricingService: motor de márgenes con resolución VARIANT→PRODUCT→CATEGORY→GLOBAL
- PriceRuleController: REST /api/v1/price-rules
- DTOs + ApiMapper para reglas de precio

Backend - Product pricing integration:
- Product/ProductDetailVariant: campos retailPrice y costPrice
- ProductUseCaseImpl: aplica márgenes automáticamente al listar/buscar productos
- Mappers actualizados para retail/cost price

+++ src/main/java/com/backandwhite/domain/model/ProductFacets.java
@@ -2,6 +2,7 @@ package com.backandwhite.domain.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,8 +16,8 @@ public class ProductFacets {
     @Builder.Default
     private List<FacetBrand> brands = new ArrayList<>();
 
-    private Double priceMin;
-    private Double priceMax;
+    private BigDecimal priceMin;
+    private BigDecimal priceMax;
 
     @Data
     @With
